Refactor: Moves YralAsyncImage to design system (#508)
This commit moves the `YralAsyncImage` composable and its related helper function `getSVGImageModel` from the `androidApp` module to the `shared/libs/designsystem` module.

This change also updates all usages of `YralAsyncImage` and `getSVGImageModel` across the `androidApp` to reference the new location in the design system module.
Additionally, Coil dependencies (`coil-compose` and `coil-svg`) are added to the `designsystem` module's build file.