Sentry Improvements and fixes (#698)
* build: Refactors Sentry KMP integration and initialization

This commit refactors the Sentry Kotlin Multiplatform (KMP) integration.

Key changes include:
*   Updating the Sentry Gradle plugin from version 4.12.0 to 5.12.2.
*   Moving Sentry KMP dependency from individual modules like `iosSharedUmbrella` to the core `shared/app` module.
*   Centralizing Sentry initialization logic into a new `initializeApp` function within the `shared/app` module, which is now called from `YralApp.kt`.
*   Disabling Sentry's auto-initialization in the Android app to allow for manual setup through the shared KMP module.
*   Configuring Sentry's Android Gradle plugin for release builds and ignoring debug variants.
*   Adding `sentry.properties` to `.gitignore`.

* feat: Adds Sentry script to upload dSYM files

Adds a new "Run Script" phase to the Xcode project for both the main and staging targets.

This script uploads the dSYM debug symbols to Sentry, which is necessary for symbolication of crash reports. It also includes a check for arm64 architecture to correctly locate the `sentry-cli` tool.